# openapi.yaml
openapi: 3.1.0
info:
  title: DeenBridge Backend API
  version: 1.0.0
  license:
    name: MIT
    url: https://github.com/Deen-Bridge/dnb-backend/blob/main/LICENSE
  description: |
    REST API for DeenBridge: authentication and sessions, courses, the book library,
    community spaces, reels, search, Stellar USDC payments, refunds and disputes,
    educator payouts, media upload signing and the SEP-1 metadata document.

    Provenance of this document:
      - Paths, methods and auth requirements are derived from app.js and src/routes/**,
        which are the authoritative mounting and middleware sources.
      - Component schemas are derived from the Mongoose schemas in src/models/**.
      - Per-endpoint request bodies are described from route middleware and the model
        fields they write. Where a controller accepts more than the model exposes, the
        body is intentionally permissive rather than invented.

    Known response inconsistencies (see issue about response standardization):
      - Most handlers return an envelope of the form { success, message, data }.
      - Some list handlers return a bare array with no envelope.
      - POST /api/email returns { success, otp, message } on success but a bare
        { message } on validation failure.
      - GET /admin/jobs returns HTML when the Accept header prefers HTML, otherwise
        { success, jobs }.
      - GET /metrics returns Prometheus text, and the SEP-1 document returns text/toml.
...
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by POST /api/auth/login or POST /api/auth/refresh.
    jobsToken:
      type: http
      scheme: bearer
      description: Static token from the JOBS_DASHBOARD_TOKEN environment variable.
...
  responses:
    Ok:
      description: Success envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SuccessResponse" }
...
    Unauthorized:
      description: Missing, malformed or expired credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    TooManyRequests:
      description: Rate limit exceeded. /api is rate limited globally and /api/auth more strictly.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      description: Standard error shape produced by the central error handler
      properties:
        success: { type: boolean, examples: [false] }
        message: { type: string }
        data: { description: Null on error responses }
        stack: { type: string, description: Present only outside production }
    SuccessResponse:
      type: object
      description: Standard success envelope. Some list handlers return a bare array instead.
      properties:
        success: { type: boolean, examples: [true] }
        message: { type: string }
        data: { description: Payload, shape depends on the endpoint }
...
paths:
...
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate the session and issue a new access token
      description: Guarded by a dedicated refresh rate limiter.
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: May instead be supplied by cookie, since cookie-parser is enabled.
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
      x-envelope-unverified: true
...